Replace INT with BIGINT

+++ jitter-rest/src/main/resources/db/migration/V1.3__create_tags_table.sql
@@ -2,9 +2,9 @@ DROP TABLE IF EXISTS `Litter`.`tags`;
 
 CREATE TABLE IF NOT EXISTS `Litter`.`tags`
 (
-    `id`         INT UNSIGNED NOT NULL AUTO_INCREMENT,
-    `tag_name`   VARCHAR(20)  NOT NULL,
-    `created_at` TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `id`         BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    `tag_name`   VARCHAR(20)     NOT NULL,
+    `created_at` TIMESTAMP       NOT NULL DEFAULT CURRENT_TIMESTAMP,
 
     PRIMARY KEY (`id`),
     UNIQUE INDEX `tag_id_UNIQUE` (`id` ASC) VISIBLE,
